Fix bugs.

+++ esp32/main/nvs.h
@@ -6,7 +6,7 @@ struct calibration_data
 {
     uint8_t motor;
     uint8_t reverse;
-    uint16_t steps;
+    double steps;
 };
 
 void init_nvs();
@@ -16,7 +16,7 @@ wifi_creds_t get_wifi_creds();
 void clear_wifi_credentials();
 void add_wifi_credentials(const char* ssid, const char* password);
 
-void set_calibration(int motor, int reverse, int steps);
+void set_calibration(int motor, int reverse, double steps);
 
 const calibration_data& get_calibration(int motor);
